% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{affs}
\alias{affs}
\title{Compute Affinity Matrix}
\usage{
affs(x, na.action = na.omit)
}
\arguments{
\item{x}{A numeric matrix (samples as rows, features as columns).}

\item{na.action}{Action to handle missing values. Defaults to \code{na.omit}.}
}
\value{
A symmetric affinity matrix.
}
\description{
Calculates the affinity matrix using Euclidean distances with local scaling.
}
