% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{es.num.graph}
\alias{es.num.graph}
\title{Estimate Optimal Number of Clusters}
\usage{
es.num.graph(W, NUMC = 2:8)
}
\arguments{
\item{W}{The affinity graph matrix.}

\item{NUMC}{A vector of possible numbers of clusters. Defaults to 2:8.}
}
\value{
A list containing the first best (\code{first}), second best (\code{second}), and all quality values.
}
\description{
Estimates the optimal number of clusters for a given graph using rotation cost.
}
