% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_time_event}
\alias{generate_time_event}
\title{Generate Time-Event Matrix}
\usage{
generate_time_event(clinical, limits, labels = NULL)
}
\arguments{
\item{clinical}{A survival matrix or data frame.}

\item{limits}{A vector of time limits.}

\item{labels}{Labels for the output columns.}
}
\description{
Creates binary event indicators at specified time limits.
}
