% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_KM}
\alias{plot_KM}
\title{Plot Kaplan-Meier Survival Curves}
\usage{
plot_KM(
  clinical,
  labels,
  limit = NULL,
  annot = NULL,
  color = NULL,
  xlab = "Follow up",
  ylab = "Survival Probability",
  title = NULL,
  legend.pos = "top",
  palette = "jama_classic",
  risk.table = T,
  risk.table.ratio = 0.4,
  anno.pos = "bottom",
  anno.x.shift = 0.5
)
}
\arguments{
\item{clinical}{Data frame containing survival data.}

\item{labels}{Vector of cluster labels.}

\item{limit}{Time limit for follow-up.}

\item{annot}{Annotation labels.}

\item{color}{Custom color palette.}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{title}{Plot title.}

\item{legend.pos}{Legend position.}

\item{palette}{Color palette.}

\item{risk.table}{Whether to show risk table.}

\item{risk.table.ratio}{Ratio of risk table height.}

\item{anno.pos}{Position of p-value annotations.}

\item{anno.x.shift}{X-axis shift for p-value annotations.}
}
\description{
Generates Kaplan-Meier survival curves based on clustering results.
}
