% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ossp_core.R
\name{plot_ossp_km}
\alias{plot_ossp_km}
\title{Plot Kaplan-Meier Survival Curves}
\usage{
plot_ossp_km(survival_data, labels)
}
\arguments{
\item{survival_data}{A data frame containing survival metadata with 'Survival' and 'Death' columns.}

\item{labels}{A vector of cluster labels.}
}
\value{
A survival plot object.
}
\description{
Generates Kaplan-Meier survival plots based on OSSP clustering results.
}
