% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{self.diffusion}
\alias{self.diffusion}
\title{Self-Diffusion on Affinity Graph}
\usage{
self.diffusion(A, K)
}
\arguments{
\item{A}{Affinity graph matrix.}

\item{K}{Number of nearest neighbors for the dominate set.}
}
\value{
A diffused symmetric matrix.
}
\description{
Implements self-diffusion logic to enhance local scaling affinity.
}
