% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spec.clu}
\alias{spec.clu}
\title{Spectral Clustering}
\usage{
spec.clu(affinity, K, type = 3)
}
\arguments{
\item{affinity}{The input affinity graph.}

\item{K}{Number of clusters.}

\item{type}{Type of Laplacian; 1: Unnormalized, 2: Random Walk, 3: Symmetric (Default).}
}
\value{
A vector of cluster labels.
}
\description{
Performs spectral clustering on an affinity matrix.
}
